+++ crates/lintel-check/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.19](https://github.com/lintel-rs/lintel/compare/lintel-check-v0.0.18...lintel-check-v0.0.19) - 2026-03-02
+
+### Other
+
+- Merge origin/master into git-integration
+- consolidate file discovery into lintel-config
+
 ## [0.0.18](https://github.com/lintel-rs/lintel/compare/lintel-check-v0.0.17...lintel-check-v0.0.18) - 2026-03-01
 
 ### Other