+++ crates/lintel-format/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.7](https://github.com/lintel-rs/lintel/compare/lintel-format-v0.0.6...lintel-format-v0.0.7) - 2026-03-02
+
+### Other
+
+- Merge origin/master into git-integration
+- consolidate file discovery into lintel-config
+
 ## [0.0.6](https://github.com/lintel-rs/lintel/compare/lintel-format-v0.0.5...lintel-format-v0.0.6) - 2026-03-01
 
 ### Other