+++ crates/lintel-validation-cache/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.10](https://github.com/lintel-rs/lintel/compare/lintel-validation-cache-v0.0.9...lintel-validation-cache-v0.0.10) - 2026-03-02
+
+### Added
+
+- use granular validation error codes like lintel::validation::required
+
 ## [0.0.9](https://github.com/lintel-rs/lintel/compare/lintel-validation-cache-v0.0.8...lintel-validation-cache-v0.0.9) - 2026-03-01
 
 ### Other